% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{completedata}
\alias{completedata}
\title{Example dataset: completedata}
\format{
A data frame with [number of rows] rows and [number of columns] variables:
\describe{
  \item{Depression}{Score of depression level in stroke patients, continuous.}
  \item{Fatalism}{Score of depression level in stroke patients, continuous.}
  \item{Comorbidity1}{number of comorbidity in stroke patients,discrete.}
  ...
}
}
\source{
Imported from \code{data/completedata.sas7bdat}.
}
\usage{
completedata
}
\description{
This dataset contains variables related to [description of dataset].
}
\keyword{datasets}
