% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summary.R
\name{generate_full_summary}
\alias{generate_full_summary}
\title{Generate Full Model Summary}
\usage{
generate_full_summary(fit_result, eval_result, response_var, predictor_vars)
}
\arguments{
\item{fit_result}{The result of a fitted linear model.}

\item{eval_result}{A list containing model evaluation metrics (R-squared, F-statistic, etc.).}

\item{response_var}{The name of the response variable.}

\item{predictor_vars}{A vector of predictor variable names.}
}
\value{
A printed model summary with all relevant statistics.
}
\description{
This function generates a full regression summary, including:
- Regression equation
- Coefficient table
- Model fit statistics
- Model evaluation
}
