% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{a formula with the form 'response~covariate+covariate...'.}

\item{data}{a data frame containing the variables specified in 'formula'.}
}
\value{
returns a list with the following elements:
\item{call}{the function that was called}
\item{coefficients}{the coefficients estimates for each predictior variable specified in model}
\item{residuals}{a vector of residuals, that is response minus fitted values, with each residual specified by observation}
\item{rank}{the numeric rank for the model}
\item{fitted.values}{the fitted values}
\item{df.residual}{the residual degrees of freedom, that is number of observations minus number of predictor variables}
\item{terms}{the terms objects that were used in model}
}
\description{
Fits univariate or multivariate linear regression models
}
\examples{
linreg(Temp~Wind+Month+Ozone,airquality)


}
