% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_sum.R
\name{linreg_sum}
\alias{linreg_sum}
\title{Linear Regression Summary}
\usage{
linreg_sum(formula, data)
}
\arguments{
\item{formula}{a formula with the form 'response~covariate+covariate...'.}

\item{data}{a data frame containing the variables specified in 'formula'.}
}
\value{
returns a list with the following elements:
\item{call}{the function that was called}
\item{sum_residuals}{summary statistics of the residuals including minimum, 1st quartile, median, mean, 3rd quartile, and maximum}
\item{coefficients}{the coefficients estimates, standard error, t-statistic value, and p-value for each predictior variable specified in model}
\item{sigma}{residual standard error}
\item{r.squared}{R squared value}
\item{adj.r.squared}{adjusted R square value}
\item{fstatistic}{f-statistic value, degrees of freedom for numerator, and degrees of freedom for denominator}
\item{f.stat.pvalue}{f-statistic p-value}
\item{residuals}{a vector of residuals, that is response minus fitted values, with each residual specified by observation}
\item{df}{degrees of freedom with the first being number of non-aliased coefficients, degrees of freedom for model, and total number of coefficients}
\item{terms}{the terms objects that were used in model}
}
\description{
Gives summary results of model along with addition statistics not included with basic `linreg()` function.
}
\examples{
linreg_sum(Temp~Wind+Month+Ozone,airquality)



}
