% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_poisson.R
\name{M_poisson}
\alias{M_poisson}
\title{Poisson Regression with IRLS}
\usage{
M_poisson(formula, data)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted.}

\item{data}{A data frame containing the variables in the model.}
}
\value{
A list containing coefficients, null deviance, residual deviance,
        the number of iterations, and convergence status.
}
\description{
This function performs Poisson regression using Iteratively Reweighted Least Squares (IRLS).
}
\examples{
library(haven)
data <- data("completedata", package = "GLM.g.p")
model_poisson <- M_poisson( Comorbidity1~ Depression, data = data)
cat("\nPoisson Regression:\n")
cat("Coefficients:\n")
print(model_poisson$coefficients)
cat("Null Deviance:", model_poisson$null_deviance, "\n")
cat("Residual Deviance:", model_poisson$residual_deviance, "\n")
cat("Converged:", model_poisson$converged, "\n")

}
